+++ src/schema.rs
@@ -4715,6 +4715,7 @@ impl TaskEffectsSpec {
 pub enum TaskNetworkEffectKind {
     Broad,
     DependencyHydration,
+    ContainerImageHydration,
     IntegrationTest,
     ToolBootstrap,
 }
@@ -4724,6 +4725,7 @@ impl TaskNetworkEffectKind {
         match self {
             Self::Broad => "broad",
             Self::DependencyHydration => "dependency_hydration",
+            Self::ContainerImageHydration => "container_image_hydration",
             Self::IntegrationTest => "integration_test",
             Self::ToolBootstrap => "tool_bootstrap",
         }